requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.2
pydantic>=2.3.0
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
groq>=0.4.0
python-multipart>=0.0.6
typing-extensions>=4.7.1
